chore: rename comp
